requirements.txt
tensorflow
opencv-python
pandas
numpy
Flask
Python-IO
pillow
json
ultralytics